Solution 22

+++ 022/SolutionTwentyTwo.h
@@ -1,8 +1,12 @@
 #ifndef SOLUTIONTWENTYTWO_H
 #define SOLUTIONTWENTYTWO_H
 
-int solutionTwentyTwo();
+long long solutionTwentyTwo();
 
 #include "../GenericFunctions.h"
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
 
 #endif SOLUTIONTWENTYTWO_H
